+++ packages/components/src/Calendar/MonthListBox.tsx
@@ -1,6 +1,8 @@
 import { Dispatch, Key, SetStateAction, useContext } from 'react';
 import { CalendarStateContext } from 'react-aria-components';
-import { Button } from '../Button';
+import { Button } from 'react-aria-components';
+import { cn } from '@marigold/system';
+import { useCalendarContext } from './Context';
 import { useFormattedMonths } from './useFormattedMonths';
 
 interface MonthDropdownProps {
@@ -17,25 +19,29 @@ const MonthListBox = ({ setSelectedDropdown }: MonthDropdownProps) => {
     state.setFocusedDate(date);
   };
 
+  const { classNames } = useCalendarContext();
+
   return (
     <ul
       data-testid="monthOptions"
-      className="grid h-full max-h-[300px] min-w-[300px] grid-cols-3 gap-y-10 overflow-y-scroll p-2"
+      className="grid h-full max-h-[300px] min-w-[300px] grid-cols-3 gap-y-10 p-2"
     >
       {months.map((month, index) => {
+        const isSelected = index === state.focusedDate.month - 1;
         return (
           <li className="flex justify-center" key={index}>
             <Button
               slot="previous"
-              variant={
-                index === state.focusedDate.month - 1 ? 'secondary' : 'text'
-              }
-              size="small"
               onPress={() => {
                 onChange(index);
                 setSelectedDropdown(undefined);
               }}
               key={index + 1}
+              aria-current={isSelected}
+              className={cn(
+                classNames.calendarListboxButton,
+                'inline-flex items-center justify-center gap-[0.5ch]'
+              )}
             >
               {month.substring(0, 3)}
             </Button>