+++ themes/theme-rui/src/components/Link.styles.ts
@@ -1,13 +1,16 @@
 import { ThemeComponent, cva } from '@marigold/system';
 
 export const Link: ThemeComponent<'Link'> = cva(
-  'underline aria-[disabled]:cursor-not-allowed py-2 underline-offset-4',
+  ' text-link font-normal aria-[disabled]:cursor-not-allowed',
   {
     variants: {
       variant: {
-        menuItemLink: '',
-        secondary: '',
+        default: '',
+        secondary: 'font-medium text-foreground underline hover:no-underline',
       },
+    },
+    defaultVariants: {
+      variant: 'default',
     },
   }
 );