+++ themes/theme-rui/src/components/Table.styles.ts
@@ -29,6 +29,7 @@ export const Table: ThemeComponent<'Table'> = {
       variant: {
         default: '',
         grid: '[&>:not(:last-child)]:border-r [&>:not(:last-child)]:border-border',
+        muted: 'bg-muted',
       },
     },
     defaultVariants: {
@@ -37,7 +38,7 @@ export const Table: ThemeComponent<'Table'> = {
   }),
   header: cva(
     [
-      'h-12 px-3 align-middle font-medium text-muted-foreground last:text-right',
+      'h-12 px-3 align-middle font-medium text-muted-foreground',
       'focus-visible:outline-2 outline-offset-2 outline-ring/70',
     ],
     {
@@ -45,6 +46,7 @@ export const Table: ThemeComponent<'Table'> = {
         variant: {
           default: '[&:has([type=checkbox])]:pr-0',
           grid: '',
+          muted: 'border-t',
         },
       },
       defaultVariants: {
@@ -72,10 +74,7 @@ export const Table: ThemeComponent<'Table'> = {
     }
   ),
   cell: cva(
-    [
-      'p-3 align-middle last:text-right',
-      'focus-visible:outline-2 outline-offset-2 outline-ring/70',
-    ],
+    ['p-3', 'focus-visible:outline-2 outline-offset-2 outline-ring/70'],
     {
       variants: {
         variant: {