+++ packages/types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @marigold/types
 
+## 1.3.1
+
+### Patch Changes
+
+- 0bca5d8: Update React aria components
+
 ## 1.3.0
 
 ### Minor Changes